+++ packages/CddVaultLink/src/package.g.ts
@@ -127,7 +127,7 @@ export async function cDDVaultSearchAsync(vaultId: number, structure?: string, s
   return await PackageFunctions.cDDVaultSearchAsync(vaultId, structure, structure_search_type, structure_similarity_threshold, protocol, run);
 }
 
-//name: CDD Vault search
+//name: CDD Vault search 2
 //input: int vaultId { nullable: true }
 //input: string molecules { category: General; nullable: true; description: Comma separated list of ids }
 //input: string names { category: General; nullable: true; description: Comma separated list of names/synonyms }
@@ -156,21 +156,22 @@ export async function cDDVaultSearchAsync(vaultId: number, structure?: string, s
 //output: dataframe result
 //meta.cache: all
 //meta.cache.invalidateOn: 0 0 * * *
-export async function cDDVaultSearch(vaultId: number, molecules: string, names: string, include_original_structures: boolean, only_ids: boolean, only_batch_ids: boolean, created_before: string, created_after: string, modified_before: string, modified_after: string, batch_created_before: string, batch_created_after: string, batch_field_before_name: string, batch_field_before_date: string, batch_field_after_name: string, batch_field_after_date: string, projects: string, data_sets: string, structure: string, structure_search_type: any, structure_similarity_threshold: number, inchikey: string, molecule_fields: string[], batch_fields: string[], fields_search: string[]) : Promise<any> {
-  return await PackageFunctions.cDDVaultSearch(vaultId, molecules, names, include_original_structures, only_ids, only_batch_ids, created_before, created_after, modified_before, modified_after, batch_created_before, batch_created_after, batch_field_before_name, batch_field_before_date, batch_field_after_name, batch_field_after_date, projects, data_sets, structure, structure_search_type, structure_similarity_threshold, inchikey, molecule_fields, batch_fields, fields_search);
+export async function cDDVaultSearch2(vaultId: number, molecules: string, names: string, include_original_structures: boolean, only_ids: boolean, only_batch_ids: boolean, created_before: string, created_after: string, modified_before: string, modified_after: string, batch_created_before: string, batch_created_after: string, batch_field_before_name: string, batch_field_before_date: string, batch_field_after_name: string, batch_field_after_date: string, projects: string, data_sets: string, structure: string, structure_search_type: any, structure_similarity_threshold: number, inchikey: string, molecule_fields: string[], batch_fields: string[], fields_search: string[]) : Promise<any> {
+  return await PackageFunctions.cDDVaultSearch2(vaultId, molecules, names, include_original_structures, only_ids, only_batch_ids, created_before, created_after, modified_before, modified_after, batch_created_before, batch_created_after, batch_field_before_name, batch_field_before_date, batch_field_after_name, batch_field_after_date, projects, data_sets, structure, structure_search_type, structure_similarity_threshold, inchikey, molecule_fields, batch_fields, fields_search);
 }
 
-//name: CDD Vault search 2
+//name: CDD Vault search
 //input: int vaultId { nullable: true }
 //input: string structure { category: Structure; nullable: true; semType: Molecule; description: SMILES,cxsmiles or mol string }
 //input: string structure_search_type { category: Structure; nullable: true; choices: ["exact","similarity","substructure"]; description: SMILES,cxsmiles or mol string }
 //input: double structure_similarity_threshold { category: Structure; nullable: true; description: A number between 0 and 1 }
 //input: int protocol { category: Protocol; nullable: true; description: Protocol id }
 //input: int run { category: Protocol; nullable: true; description: Specific run id }
+//input: int page_size { nullable: true; description: Page size for preview (defaults to PREVIEW_ROW_NUM if omitted) }
 //output: dataframe result
 //meta.cache: all
 //meta.cache.invalidateOn: 0 0 * * *
 //editor: Cddvaultlink:CDDVaultSearchEditor
-export async function cDDVaultSearch2(vaultId: number, structure?: string, structure_search_type?: any, structure_similarity_threshold?: number, protocol?: number, run?: number) : Promise<any> {
-  return await PackageFunctions.cDDVaultSearch2(vaultId, structure, structure_search_type, structure_similarity_threshold, protocol, run);
+export async function cDDVaultSearch(vaultId: number, structure?: string, structure_search_type?: any, structure_similarity_threshold?: number, protocol?: number, run?: number, page_size?: number) : Promise<any> {
+  return await PackageFunctions.cDDVaultSearch(vaultId, structure, structure_search_type, structure_similarity_threshold, protocol, run, page_size);
 }